requirements.txt
txt

langchain
faiss-cpu
sentence-transformers
huggingface-hub
gradio
datasets
python-dotenv